refactor yatai service integration test
